Moved KWKekt into a new module flixCompsExperimental.py
